Support custom 'joiner' instead of comma
This commit adds support for a custom 'joiner' instead of a comma.
For example,

   {2$$and$$opt1|opt2|opt3}

will join together 2 selections from opt1,opt2,opt3 with the word
'and' instead of a comma, eg

  opt3 and opt1

Omitting the second $$ falls back to the default ','

Additionally, if a larger quantity is selected than number of
options provided then options will be selected with replacement.
This allows you to pull multiple samples from a wildcard file, eg

 {3$$featuring$$__artist__}

will take 3 random entries from artist.txt and join them with 'featuring'