Updated dependencies
dynamicprompts is now always upgraded if a new version is available. This means that a new version of the extension does not need to be released every time dynamicprompts updates.
The new version 0.1.18 of dynamicprompts removes an explicit dependency on pytorch and relies on the version that is installed with the webui. This avoids an unpleasant 1.8GB download.
Unpinned requests so that it uses the same version as Automatic1111.